third_party/blink/web_tests/fast/xsl/xslt-processor-template-expected.txt

CONSOLE WARNING: Usage of XSLTProcessor or XSLT Processing Instructions was detected. These features have been deprecated by all browsers, and a special early trial of complete removal is underway in this browser.
--> If you are a *user* experiencing a problem, please report the issue directly to the operator of the website.
--> If you are a site owner, and you think this trial is causing an unexpected issue, please report a bug at https://issues.chromium.org/issues/new?component=1456730&template=2210866
CONSOLE ERROR: Uncaught NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled
This tests that XSLT transforms can traverse into XHTML template element content when applying XSL template. If the test succeeds, the transform will have swapped the position of the body spans (A and B) with the template content spans (C and D) and replaced the spans with divs.
